+++ R/here_read.R
@@ -1,6 +1,7 @@
-#' Read Data Frame or Object from RDS File in a Specified Directory
+#' Read an R Object from RDS in a Specified Directory
 #'
-#' Reads an RDS file specified by `name` from a directory defined by `dir_path` or `push_mods`, returning the data frame or object stored within.
+#' Reads an RDS file specified by `name` from a directory defined by `dir_path`
+#' or `push_mods`, returning the R object stored within.
 #' If no .rds file is found, it searches for the file without the .rds extension.
 #' This function uses the `here` package to resolve the path, ensuring that file paths are built in a consistent and platform-independent manner.
 #'
@@ -12,16 +13,16 @@
 #' If `dir_path` is NULL, the `push_mods` variable must be defined in the user's environment or within the package, pointing to the directory from where files are to be read.
 #' This function will first try to read an .rds file. If not found, it will attempt to read the file without the .rds extension.
 #'
-#' @return The data frame or R object stored in the file.
+#' @return The R object stored in the file.
 #'
 #' @examples
 #' \dontrun{
 #' # Assuming `push_mods` is set in your environment to "~/mydata"
-#' # and you have previously saved an RDS file named "my_df.rds" in that directory
-#' my_df <- here_read("my_df")
+#' # and you have previously saved an RDS file named "result.rds" in that directory
+#' result <- here_read("result")
 #'
 #' # Reading from a custom directory
-#' my_df <- here_read("my_df", dir_path = "~/custom_dir")
+#' result <- here_read("result", dir_path = "~/custom_dir")
 #' }
 #'
 #' @export