+++ R/margot_plot_boxplot_covariate.R
@@ -258,7 +258,7 @@ margot_plot_boxplot_covariate <- function(data,
           dpi = 300
         )
 
-        margot::here_save_qs(p, filename, save_path, preset = "high", nthreads = 1)
+        margot::here_save(p, filename, dir_path = save_path)
 
         cli::cli_alert_success("Plot saved successfully")
       } else {