+++ R/margot_plot_histogram.R
@@ -309,7 +309,7 @@ margot_plot_histogram <- function(data,
           dpi = 500
         )
 
-        margot::here_save_qs(p, filename, save_path, preset = "high", nthreads = 1)
+        margot::here_save(p, filename, dir_path = save_path)
 
         cli::cli_alert_success("Plot saved successfully as '{filename}' in '{save_path}'")
       } else {